Implement PalHijack for Unix
This change implements PalHijack for Unix. To ensure that the target
thread doesn't progress while it is being hijacked, it is suspended
on OSX or an activation signal is injected and the thread is held
waiting in the signal handler on other Unixes.

Since it is not safe to use pthread mutex in a signal handler for
synchronization between the thread performing the hijack and the
signal handler, the implementation uses eventfd on Linux or pipe on
other Unixes where the better performing eventfd doesn't exist.

It is not safe to use pthread mutex in signal handler.